code/requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# Vision + Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# Utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.66.0

# For DINOv2 backbone via torch.hub
timm>=0.9.0
